Re-Creates entire website using bootstrap
This update recodes the entire website to use bootstrap 3.0 in order to
make the website responsive and look nicer. Updates the color scheme
to a green/black/grey theme. Removes the “contact” navigation bar
item and instead adds a universal footer that has relevant information
to contact myself. Adds an “About Me” section to the home page to
give an elevator pitch about my professional self to anyone who sees it.
Portfolio page is created that links to my github account and also hosts
two web applications I coded. The first app is an Overwatch hero counter
that displays counter-picks to all heros in the Overwatch videogame.
The second app is an article summarizer based on SMMRY.com. Creates
a rudimentary blog engine and stores first blog article.